+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Broken API: change ReadTimeout/WriteTimeout to IdleTimeout
 - Support passing Deadline of client contexts to server side
 - remove InprocessClient plugin
+- use heartbeat/tcp_keepalive to avoid client hanging
 
 
 ## 5.0